+++ src/Task/BatchProgressTask.php
@@ -67,6 +67,9 @@ abstract class BatchProgressTask extends BaseTask
                 try {
                     $taskResult = $this->processItem($item);
                 } catch (Throwable $e) {
+                    $this->logger->error($e->getMessage(), array_merge($this->getLogContext(), [
+                        'exception' => $e
+                    ]));
                     $taskResult = new ErrorResult($e->getMessage(), [], $e);
                 }